+++ tests/App/Screen/TestOrchidLogListScreen.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Screen;
+namespace Tests\App\Screen;
 
 use Czernika\OrchidLogViewer\Screen\OrchidLogListScreen;