+++ tests/TestCase.php
@@ -2,24 +2,16 @@
 
 namespace Tests;
 
-use App\Models\User;
 use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Orchid\Support\Facades\Alert;
 use Orchid\Support\Facades\Dashboard;
-use Orchid\Support\Testing\ScreenTesting;
-use Plannr\Laravel\FastRefreshDatabase\Traits\FastRefreshDatabase;
 use Tabuna\Breadcrumbs\Breadcrumbs;
 use Watson\Active\Active;
 
 abstract class TestCase extends BaseTestCase
 {
-    use FastRefreshDatabase, ScreenTesting, WithWorkbench;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-    }
+    use WithWorkbench;
 
     /**
      * @param  \Illuminate\Foundation\Application  $app
@@ -40,10 +32,5 @@ abstract class TestCase extends BaseTestCase
             'Breadcrumbs' => Breadcrumbs::class,
             'Dashboard' => Dashboard::class,
         ];
-    }
-
-    public function admin()
-    {
-        return User::factory()->admin()->create();
     }
 }