+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Releases
 
+## v1.0.1 Jan 11, 2024 
+
+- Fix: fixing issue when log route name in breadcrumbs was ignoring config value
+- Feat: added separate options for menu localization
+- Tests: added tests for default log screen
+
 ## v1.0.0 Jan 10, 2024 
 
 - Initial release